+++ app/src/main/java/org/localsend/miuix/LocalSendApplication.kt
@@ -1,0 +1,11 @@
+package org.localsend.miuix
+
+import android.app.Application
+import org.localsend.miuix.network.TlsStore
+
+class LocalSendApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        TlsStore.init(this)
+    }
+}